requirements.txt
streamlit
langchain
PyPDF2
python-pptx
fastapi
uvicorn
requests
huggingface_hub
transformers
diffusers
gradio
langchain_google_genai
langchain_groq
ctransformers
pandas
numpy
matplotlib
torch
google-generativeai
groq
langchain_community
langchain_core
pandasai
mesop